Fix errors in tui crate

+++ crates/monte-cardo-core/src/consts.rs
@@ -9,8 +9,8 @@ pub const MAX_CARD_NUMBER: usize = 18;
 
 // Default deck for "The Great Dalmuti"
 // The first position is always wilds, or "jesters"
-pub const DEFAULT_DALMUTI_DECK: [u16; 18] =
+pub const DEFAULT_DALMUTI_DECK: [usize; 18] =
     [2, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 0, 0];
 
 // Default deck for "Scum"
-pub const DEFAULT_SCUM_DECK: [u16; 18] = [2, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0];
+pub const DEFAULT_SCUM_DECK: [usize; 18] = [2, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0];
